Update LICENCE
Switched to ZLib, because then this engine would not:

1. Be able to be used on consoles (if it ever goes to that level, I doubt it, but who knows) and publishers would not even accept it.
2. Be able to even USE PhysX, I would go on Bullet, but I don't want to sacrifice some features (until said otherwise)

Now confusion still is around, but now even less. Qt is a LGPL library, and in the website it says that the software must be compliant with either GPL, LGPLv2.1, GPLv3 or LGPLv3 licenses. I feel a bit relieved, but if there are some problems with this, pull an issue about it, maybe you can make me reconsider or recommend me something better.