Finally get Ogre to render, but...
Can't load meshes or, pretty much anything because it throws file not
found errors on the resource groups when the resource groups were
actually loaded correctly, and when the files in said groups do exist.

Right now it shows a nice light blue color as the background of the
render (which is the entire screen really). You can change this in
oBgColor on qogrewindow.cpp on line 21.